Renamed and moved the "GitHubController" class (since it's not really a controller)
